Move Printers that write to logs to jvm package
